gnome-desktop-defaults: Removing tilingshell from defaults
Firstly, I'm *personally* a big fan of the tilingshell extension
but we're also not going to forcibly put too much weight on upstream
developers for being the default, ie the recent dash to panel stuff.

So, lets lighten the level of customisation to the bare essentials and
make each GNOME stack upgrade easier, whilst hopefully our users will still
use the packaged builds and donate to the authors appropriately. There is also
the argument that by bundling by default, we may be removing the discovery
of revenue/support options for those developers. Being indie devs ourselves,
we know what it is to struggle.

Signed-off-by: Ikey Doherty <ikey@aerynos.com>